Move object files component to compiling capability
