+++ subprojects/platform-native/src/main/java/dev/nokee/platform/nativebase/internal/compiling/ObjectFiles.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 the original author or authors.
+ * Copyright 2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.platform.nativebase.internal;
+package dev.nokee.platform.nativebase.internal.compiling;
 
 import dev.nokee.model.internal.core.ModelComponent;
 import org.gradle.api.provider.Provider;
